LPS-67552 Rename package
